requirements.txt
Django==3.2.11
django-oauth2-provider
djangorestframework